Here's a proposed commit message for the changes you mentioned:
In this commit message, we're using the  type to indicate that we've made changes to the feature or functionality of our project. The  scope indicates that we're making changes related to the number of lines added in the README file.
The message itself describes the changes made and why they were made. We've updated the number of lines added to 9711 from 9708, which is a breaking change compared to the previous value. Additionally, we've also updated the active time to 57 minutes from 54 minutes.
The  section is included at the end of the message and highlights any breaking changes introduced in this commit. In this case, there are two breaking changes:
1. The number of lines added has been updated to 9711, which is a breaking change compared to the previous value of 9708.
2. The active time has been updated to 57 minutes from 54 minutes.
By following conventional commit best practices, we make it easier for others to understand and navigate our project's history.