The commit message for this git diff could be:
In this message, the type is  to indicate that the changes made were not related to new features or bug fixes, but rather an improvement to existing code. The scope is , as the changes made were specific to this project.
The summary of the commit message explains the change and why it was made. In this case, the changes were made to improve the accuracy of the coding activity statistics.
The modified files list includes the files that were changed in the commit, with the  symbol indicating files that were removed and the  symbol indicating files that were added.