Here is a concise commit message based on the given git diff and modified files:
This commit message uses the conventional commit type  to indicate that the changes made are related to maintaining or managing the project, rather than adding new functionality or fixing a bug. The message also includes a brief summary of what changed and why, in this case, updating the coding activity report for 03/02/2025.