Here's a possible commit message based on the provided information:
This commit message uses the conventional commit type , which indicates that we have made changes to our codebase to fix a problem or issue. The scope of the change is the  project, and the message provides a brief summary of the changes made, including the correction of errors in  and , as well as an update to the visualizations and last updated timestamp.
The message also includes a list of modified files:  and . This information helps other developers understand which parts of the codebase were affected by this commit, and what changes were made to those files.