Here's a possible commit message based on the provided git diff:
In this message, we use the conventional commit type "refactor" to indicate that we have improved the code without adding any new features or fixing any bugs. We also describe the changes we have made in detail, including the files that have been modified and the reasons for our optimization choices.