Here is a concise commit message based on the given Git diff and modified files:
This message follows the conventional commit format and indicates that we are refactoring our code to improve the coding activity visualization. The changes made include adding more detailed data to the visualizations, which will provide a more accurate representation of our coding activity. This commit message is clear and descriptive, and it provides enough information for others to understand what has changed in the codebase.