Here is a concise commit message based on the provided git diff and modified files:
In this message, we've focused on the main changes and affected functionality. We've used conventional commit types (chore) for the commit message to indicate that it's a change related to maintaining the project, rather than adding new features or fixing bugs. We've also provided a brief summary of the changes made in each file, including the number of lines added and removed and the total active time spent on the files. Finally, we've included the duration of the activity in minutes.