Here is a revised commit message that adheres to the conventional commit format:
In this message, we use the conventional commit type "feat" to indicate that this is a new feature that was added. We also include a brief description of the changes made in the commit, which helps other developers understand what this commit does and why it was necessary. Finally, we list the modified files that were affected by this commit, which is helpful for tracking down bugs and ensuring that all changes are properly recorded in the version control system.