The diff between the previous version of the  file and the current version shows that there have been several modifications to the file, including the addition of new entries for the modified files, the updating of the line count statistics, and the addition of a new section describing the modified files. The message in the commit is also updated to reflect the changes made in the file. Here's an example of how the diff might look like:
In this example, the diff shows that there have been several changes made to the  file between the previous version and the current version. The diff includes the addition of a new entry for the modified files, as well as an update to the line count statistics and the addition of a new section describing the modified files. The message in the commit is also updated to reflect the changes made in the file.