Here's a concise commit message based on the given git diff and modified files:
This message follows the conventional commit format of including a type (in this case, ) and a brief summary. The summary describes the change in a way that is easy to understand for anyone reading the commit history. In this case, we're updating the total lines added and removed in the activity file and adding a new file.
The message also includes details about the affected files ( and ) and the duration of the activity (16 minutes).