This is an example of a concise commit message based on the provided git diff and modified files:
In this commit message, we have summarized the main changes in a concise and descriptive way. We have used conventional commit types (e.g., "feat" for new features, "fix" for bug fixes) to indicate the purpose of the commit. The message also includes the names of the modified functions, files, and the duration of the activity.