Here's a possible commit message based on the information provided:
This commit message clearly describes the changes made to the coding activity statistics for the CDA project in 2025-02-24. It includes information about the modified functions and files, as well as the last updated time. The use of conventional commit types (refactor) helps ensure that the commit message is understandable and easy to read for other developers who may be reviewing the changes.