Here is a possible commit message based on the provided git diff:
This commit message uses the conventional commit type "refactor" to indicate that changes were made to existing code in a way that does not add new features or fix bugs, but rather improves its performance or maintainability. The message also includes a brief description of the changes made (updating activity statistics for Feb 25th), as well as a list of the affected files and the duration of the activity.