Here is a concise commit message based on the provided git diff and modified files:
This commit message provides an overview of the changes made in the coding activity for 2025-03-04, including updates to total lines added, removed, and active time, as well as new files and activities added or removed. The message also includes a list of affected files and the duration of the activity.