Here is a suggested commit message for this change:
This message provides a clear description of the changes made in the last 24 hours, including the number of lines added and removed, as well as the net change and active time. It also mentions that the activity statistics have been updated in the README file and that two files have been modified.